use persistent state instead of separate packages state
